XP-4625 ContentSelector- Pressing edit-button does not open the new browser tab with a content-wizard page -Updated wizard tab to listen to 'edit' event -Refactored
